requirements.txt
pandas
numpy
matplotlib
seaborn
Pyyaml
tqdm
pydantic